--- sources/Google.Solutions.Ssh.Test/Native/SshAssert.cs
@@ -1,31 +1,0 @@
-﻿using Google.Solutions.Common.Util;
-using Google.Solutions.Ssh.Native;
-using NUnit.Framework;
-using System;
-using System.Threading.Tasks;
-
-namespace Google.Solutions.Ssh.Test.Native
-{
-    internal static class SshAssert
-    {
-        public static void ThrowsNativeExceptionWithError(
-            SshSession session,
-            LIBSSH2_ERROR expected,
-            Action action)
-        {
-            try
-            {
-                action();
-                Assert.Fail("Expected SshNativeException with error " + expected);
-            }
-            catch (Exception e) when (!(e is AssertionException))
-            {
-                Assert.IsInstanceOf(typeof(SshNativeException), e.Unwrap());
-                Assert.AreEqual(expected, ((SshNativeException)e.Unwrap()).ErrorCode);
-
-                Assert.IsTrue(session.LastError == LIBSSH2_ERROR.NONE ||
-                              session.LastError == expected);
-            }
-        }
-    }
-}
